# ACL/src/main/resources/static/swagger.yml
swagger: '2.0'
info:
  description: security API documentation
  version: 0.0.1
  title: security API
  contact: {}
  license:
    name: unlicensed
host: localhost:8085
basePath: /
...
paths:
...
  /api/acl-classes:
    get:
      tags:
        - acl-class-resource
      summary: getAllAclClasses
      operationId: getAllAclClassesUsingGET
      produces:
        - '*/*'
      parameters:
        - name: page
          in: query
          description: Page number of the requested page
          required: false
          type: integer
          format: int32
        - name: size
          in: query
          description: Size of a page
          required: false
          type: integer
          format: int32
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AclClassDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - acl-class-resource
      summary: createAclClass
      operationId: createAclClassUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: aclClassDTO
          description: aclClassDTO
          required: true
          schema:
            $ref: '#/definitions/AclClassDTO'
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AclClassDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/acl-classes/{id}:
    get:
      tags:
        - acl-class-resource
      summary: getAclClass
      operationId: getAclClassUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AclClassDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - acl-class-resource
      summary: updateAclClass
      operationId: updateAclClassUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: aclClassDTO
          description: aclClassDTO
          required: true
          schema:
            $ref: '#/definitions/AclClassDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AclClassDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - acl-class-resource
      summary: deleteAclClass
      operationId: deleteAclClassUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    patch:
      tags:
        - acl-class-resource
      summary: partialUpdateAclClass
      operationId: partialUpdateAclClassUsingPATCH
      consumes:
        - application/merge-patch+json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: aclClassDTO
          description: aclClassDTO
          required: true
          schema:
            $ref: '#/definitions/AclClassDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AclClassDTO'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
definitions:
  AclClassDTO:
    type: object
    required:
      - classIdType
      - className
    properties:
      classIdType:
        type: string
      className:
        type: string
      id:
        type: integer
        format: int64
    title: AclClassDTO
...